Remove dead code: pr_get_state export, misnamed test file, and common.ts sync duplication #171
