+++ package.json
@@ -1,5 +1,5 @@
 {
-	"version": "0.76.0",
+	"version": "0.77.0",
 	"name": "@joshuafolkken/kit",
 	"type": "module",
 	"license": "MIT",
